PyQt5 compat

+++ gitfourchette/widgets/qelidedlabel.py
@@ -36,7 +36,7 @@ class QElidedLabel(QLabel):
         elideMode = self.elideMode()
         metrics = self.fontMetrics()
         margin = self.margin()
-        m = metrics.horizontalAdvance('x') / 2 - margin
+        m = int(metrics.horizontalAdvance('x') / 2 - margin)  # int() for PyQt5 compat
         r = self.contentsRect().adjusted(margin + m,  margin, -(margin + m), -margin)
         width = r.width()
